Moving all scripts away from bash, and towards nodejs as the standard backdrop
